+++ app/src/main/java/com/example/chat/di/AppModule.kt
@@ -4,8 +4,6 @@ package com.example.chat.di
 import android.content.Context
 import androidx.room.Room
 import com.example.chat.data.*
-import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.firestore.FirebaseFirestore
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -28,16 +26,4 @@ object AppModule {
     @Provides
     fun provideRepository(dao: ContactDao): ContactRepository = ContactRepositoryImpl(dao)
 }
-
-    @Module
-
-    @InstallIn(SingletonComponent::class)
-    object FirebaseModule {
-
-        @Provides
-        fun provideFirebaseAuth(): FirebaseAuth = FirebaseAuth.getInstance()
-
-        @Provides
-        fun provideFirestore(): FirebaseFirestore = FirebaseFirestore.getInstance()
-    }
