+++ app/src/main/java/com/example/chat/ui/presentation/AddEmergencyContactScreen.kt
@@ -1,36 +1,17 @@
-package com.example.chat.ui
+package com.example.chat.ui.presentation
 
-import com.example.chat.viewmodel.ContactViewModel
-
-import androidx.compose.ui.Modifier
-
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Button
-import androidx.compose.material3.OutlinedTextField
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
+import androidx.compose.foundation.layout.*
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.hilt.navigation.compose.hiltViewModel
-
 import com.example.chat.data.Contact
-
+import com.example.chat.viewmodel.ContactViewModel
 
 @Composable
-fun AddEmergencyContactScreen(
-    onNavigateBack: () -> Unit,
-    viewModel: ContactViewModel = hiltViewModel()
-) {
+fun AddEmergencyContactScreen(onNavigateBack: () -> Unit, viewModel: ContactViewModel) {
     var name by remember { mutableStateOf("") }
     var phone by remember { mutableStateOf("") }
