+++ app/src/main/java/com/example/chat/ui/presentation/HomeScreen.kt
@@ -1,23 +1,21 @@
-package com.example.chat.ui
+package com.example.chat.ui.presentation
 
 import android.content.Intent
 import android.net.Uri
 import androidx.compose.foundation.layout.*
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.ArrowForward
 import androidx.compose.material.icons.filled.*
 import androidx.compose.material3.*
-import androidx.compose.runtime.Composable
+import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.navigation.NavController
+import com.example.chat.ui.view.StyledButton
 
 @Composable
 fun HomeScreen(
-    navController: NavController,
     onNavigateToAddContact: () -> Unit,
     onNavigateToContacts: () -> Unit
 ) {
@@ -49,7 +47,7 @@ fun HomeScreen(
             }) {
                 Icon(imageVector = Icons.Default.Warning, contentDescription = "SOS")
                 Spacer(modifier = Modifier.width(8.dp))
-                Text("SOS")
+                Text(text = "SOS")
             }
 
             Button(onClick = {
@@ -58,7 +56,7 @@ fun HomeScreen(
             }) {
                 Icon(imageVector = Icons.Default.LocationOn, contentDescription = "Live Location")
                 Spacer(modifier = Modifier.width(8.dp))
-                Text("Share live location")
+                Text(text = "Share live location")
             }
         }
 
@@ -70,7 +68,7 @@ fun HomeScreen(
         ) {
             Icon(imageVector = Icons.Default.AccountCircle, contentDescription = "Add Contact")
             Spacer(modifier = Modifier.width(8.dp))
-            Text("Add Emergency Contact")
+            Text(text = "Add Emergency Contact")
         }
 
         Spacer(modifier = Modifier.height(20.dp))
@@ -81,67 +79,12 @@ fun HomeScreen(
         ) {
             Icon(imageVector = Icons.Default.Call, contentDescription = "View Contacts")
             Spacer(modifier = Modifier.width(8.dp))
-            Text("View Contacts")
+            Text(text = "View Contacts")
         }
+        Spacer(modifier = Modifier.height(30.dp))
 
-        Spacer(modifier = Modifier.weight(1f))
-
-        BottomNavigationBar(navController)
-    }
-}
-
-@Composable
-fun StyledButton(text: String, query: String) {
-    val context = LocalContext.current
-    Button(
-        onClick = {
-            val intent = Intent(Intent.ACTION_VIEW, Uri.parse("geo:0,0?q=$query"))
-            context.startActivity(intent)
-        },
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(vertical = 6.dp)
-    ) {
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            horizontalArrangement = Arrangement.SpaceBetween,
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Text(text)
-            Icon(Icons.AutoMirrored.Filled.ArrowForward, contentDescription = "Arrow")
-        }
-    }
-}
-
-@Composable
-fun BottomNavigationBar(navController: NavController) {
-    NavigationBar {
-        NavigationBarItem(
-            icon = { Icon(Icons.Default.Home, contentDescription = "Home") },
-            label = { Text("Home") },
-            selected = true,
-            onClick = { navController.navigate("home") }
-        )
-        NavigationBarItem(
-            icon = { Icon(Icons.Default.LocationOn, contentDescription = "Live Location") },
-            label = { Text("Live Location") },
-            selected = false,
-            onClick = {
-                val intent = Intent(Intent.ACTION_VIEW, Uri.parse("geo:0,0?q=My+Location"))
-                navController.context.startActivity(intent)
-            }
-        )
-        NavigationBarItem(
-            icon = { Icon(Icons.Default.Call, contentDescription = "Contacts") },
-            label = { Text("Contacts") },
-            selected = false,
-            onClick = { navController.navigate("contacts") }
-        )
-        NavigationBarItem(
-            icon = { Icon(Icons.Default.Person, contentDescription = "Profile") },
-            label = { Text("Profile") },
-            selected = false,
-            onClick = { /* future screen */ }
-        )
+        StyledButton("Police station near me", "police station near me")
+        StyledButton("Hospital near me", "hospital near me")
+        StyledButton("Emergency Helpline", "emergency helpline")
     }
 }
