+++ app/src/main/java/com/example/chat/ui/presentation/ProfileForm.kt
@@ -1,4 +1,4 @@
-package com.example.chat.ui.profile
+package com.example.chat.ui.presentation
 
 import androidx.compose.foundation.layout.*
 import androidx.compose.material3.*
@@ -6,6 +6,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
+import com.example.chat.datamodel.UserProfile
 
 
 @Composable
