+++ app/src/main/java/com/example/chat/viewmodel/ProfileViewModel.kt
@@ -1,8 +1,9 @@
-package com.example.chat.ui.profile
-
+package com.example.chat.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.example.chat.datamodel.UserProfile
+import com.example.chat.repo.ProfileRepository
 import com.google.firebase.auth.FirebaseAuth
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
