+++ build.gradle.kts
@@ -3,11 +3,11 @@ plugins {
     alias(libs.plugins.android.application) apply false
     alias(libs.plugins.kotlin.android) apply false
     alias(libs.plugins.kotlin.compose) apply false
-
-    id("com.google.devtools.ksp") version "2.1.10-1.0.31" apply false
     alias(libs.plugins.google.gms.google.services) apply false
-
+    alias(libs.plugins.kapt) apply false
+    alias(libs.plugins.ksp) apply false
 
     id("com.google.dagger.hilt.android") version "2.49" apply false
 }
+
